requirements.txt
pandas==2.2.2
google-generativeai==0.7.2
python-dotenv==1.0.1
plotly==5.22.0
sqlalchemy==2.0.31
streamlit